Rollback the updated META6 template
Apparently, the docs' reference to the META6 spec was incorrect, and
nobody seems able to link me to anything resembling a more accurate
document. For now, stick to the template shown on docs.perl6.org.

Additionally, a PR has been submitted to docs.perl6.org to remove the
reference to the "incorrect spec".